unit tests and integration/e2e tests are now separate
