+++ Makefile
@@ -111,6 +111,10 @@ lint-fix:
 
 .PHONY: test
 test: manifests generate fmt vet envtest ## Run tests.
+	go test ./apis... ./controllers/... -coverprofile operator.out
+
+.PHONY: test-integration
+test-integration: manifests generate fmt vet envtest ## Run tests.
 	KUBEBUILDER_ASSETS="$(shell $(ENVTEST) use $(ENVTEST_K8S_VERSION) --bin-dir $(LOCALBIN) -p path)" go test ./... -coverprofile operator.out
 
 ##@ Build
