+++ test/integration/executor_controller_integration_test.go
@@ -1,4 +1,4 @@
-package install
+package integration
 
 import (
 	. "github.com/onsi/ginkgo/v2"
