+++ test/integration/lookout_controller_e2e_test.go
@@ -1,4 +1,4 @@
-package install
+package integration
 
 import (
 	"io"
@@ -10,18 +10,18 @@ import (
 	"github.com/armadaproject/armada-operator/controllers/utils"
 )
 
-var executorYaml = `apiVersion: install.armadaproject.io/v1alpha1
-kind: Executor
+var lookoutYaml = `apiVersion: install.armadaproject.io/v1alpha1
+kind: Lookout
 metadata:
   labels:
-    app.kubernetes.io/name: executor
-    app.kubernetes.io/instance: executor-sample
+    app.kubernetes.io/name: lookout
+    app.kubernetes.io/instance: lookout-sample
     app.kubernetes.io/part-of: armada-operator
     app.kubernetes.io/created-by: armada-operator
-  name: executor-e2e
+  name: lookout-e2e
 spec:
   image:
-    repository: test-executor
+    repository: test-lookout
     tag: latest
   applicationConfig:
     server: example.com:443
@@ -32,10 +32,10 @@ spec:
 `
 
 var _ = Describe("Armada Operator", func() {
-	When("User applies Executor YAML using kubectl", func() {
-		It("Kubernetes should create Executor Kubernetes resources", func() {
-			By("Calling the Executor Controller Reconcile function", func() {
-				f, err := utils.CreateTempFile([]byte(executorYaml))
+	When("User applies Lookout YAML using kubectl", func() {
+		It("Kubernetes should create Lookout Kubernetes resources", func() {
+			By("Calling the Lookout Controller Reconcile function", func() {
+				f, err := utils.CreateTempFile([]byte(lookoutYaml))
 				defer func() {
 					Expect(f.Close()).ToNot(HaveOccurred())
 				}()
